Split up contexts vs mapreduce logic
