+++ bin/src/server/script/mapreduce.rs
@@ -1,68 +1,36 @@
 use rlua::{Table, Function};
-use rlua::prelude::*;
 use serde_json::value::Value as JsonValue;
-use std::path::Path;
 use uuid::Uuid;
-use indradb::{Vertex, Datastore};
+use indradb::Vertex;
 use statics;
 use crossbeam_channel::{Receiver, Sender, bounded};
 use std::time::Duration;
 use std::thread::{spawn, JoinHandle};
 use super::errors;
+use super::context;
 use super::converters;
 
 const CHANNEL_CAPACITY: usize = 1000;
 const CHANNEL_RECV_TIMEOUT_SECONDS: u64 = 1;
 const REPORT_SECONDS: u64 = 30;
 
-pub fn create_lua_context(account_id: Uuid, arg: JsonValue) -> Result<Lua, errors::ScriptError> {
-    let l = Lua::new();
-
-    {
-        let globals = l.globals();
-
-        // Update the `package.path` to include the script root, so it's easier
-        // for scripts to require each other.
-        {
-            let package: Table = globals.get("package")?;
-            let old_path: String = package.get("path")?;
-            let script_path = Path::new(&statics::SCRIPT_ROOT[..])
-                .join("?.lua")
-                .to_str()
-                .unwrap()
-                .to_string();
-            package.set("path", format!("{};{}", old_path, script_path))?;
-        }
-
-        // Create a new transaction for the script
-        let trans = statics::DATASTORE.transaction(account_id)?;
-
-        // Add globals
-        globals.set("trans", converters::ProxyTransaction::new(trans))?;
-        globals.set("account_id", account_id.to_string())?;
-        globals.set("arg", converters::JsonValue::new(arg))?;
-    }
-
-    Ok(l)
-}
-
-enum MapReduceWorkerTask {
+enum WorkerTask {
     Map(Vertex),
     Reduce((converters::JsonValue, converters::JsonValue))
 }
 
-struct MapReduceWorker {
+struct Worker {
     thread: JoinHandle<Result<(), errors::MapReduceError>>,
     shutdown_sender: Sender<()>
 }
 
-impl MapReduceWorker {
-    fn start(account_id: Uuid, contents: String, path: String, arg: JsonValue, in_receiver: Receiver<MapReduceWorkerTask>, out_sender: Sender<converters::JsonValue>) -> Self {
+impl Worker {
+    fn start(account_id: Uuid, contents: String, path: String, arg: JsonValue, in_receiver: Receiver<WorkerTask>, out_sender: Sender<converters::JsonValue>) -> Self {
         let (shutdown_sender, shutdown_receiver) = bounded::<()>(1);
 
         let thread = spawn(move || -> Result<(), errors::MapReduceError> {
             let mut should_shutdown = false;
-            let l = create_lua_context(account_id, arg).map_err(|err| errors::MapReduceError::WorkerSetup(err))?;
+            let l = context::create(account_id, arg).map_err(|err| errors::MapReduceError::WorkerSetup(err))?;
             let table: Table = l.exec(&contents, Some(&path)).map_err(|err| errors::MapReduceError::WorkerSetup(errors::ScriptError::Lua(err)))?;
             let mapper: Function = table.get("map").map_err(|err| errors::MapReduceError::WorkerSetup(errors::ScriptError::Lua(err)))?;
             let reducer: Function = table.get("reduce").map_err(|err| errors::MapReduceError::WorkerSetup(errors::ScriptError::Lua(err)))?;
@@ -71,10 +39,10 @@ impl MapReduceWorker {
                 select_loop! {
                     recv(in_receiver, task) => {
                         let value = match task {
-                            MapReduceWorkerTask::Map(vertex) => {
+                            WorkerTask::Map(vertex) => {
                                 mapper.call(converters::Vertex::new(vertex)).map_err(|err| errors::MapReduceError::MapCall(err))
                             },
-                            MapReduceWorkerTask::Reduce((first, second)) => {
+                            WorkerTask::Reduce((first, second)) => {
                                 reducer.call((first, second)).map_err(|err| errors::MapReduceError::ReduceCall(err))
                             }
                         }?;
@@ -108,24 +76,24 @@ impl MapReduceWorker {
     }
 }
 
-pub struct MapReduceWorkerPool {
+pub struct WorkerPool {
     reporter_thread: JoinHandle<()>,
     router_thread: JoinHandle<Result<JsonValue, errors::MapReduceError>>,
     in_sender: Sender<Vertex>,
     shutdown_sender: Sender<()>
 }
 
-impl MapReduceWorkerPool {
+impl WorkerPool {
     pub fn start(account_id: Uuid, contents: String, path: String, arg: JsonValue) -> Self {
         let (mapreduce_in_sender, mapreduce_in_receiver) = bounded::<Vertex>(CHANNEL_CAPACITY);
-        let (worker_in_sender, worker_in_receiver) = bounded::<MapReduceWorkerTask>(CHANNEL_CAPACITY);
+        let (worker_in_sender, worker_in_receiver) = bounded::<WorkerTask>(CHANNEL_CAPACITY);
         let (worker_out_sender, worker_out_receiver) = bounded::<converters::JsonValue>(CHANNEL_CAPACITY);
         let (reporter_sender, reporter_receiver) = bounded::<()>(0);
         let (shutdown_sender, shutdown_receiver) = bounded::<()>(2);
-        let mut worker_threads: Vec<MapReduceWorker> = Vec::with_capacity(*statics::MAP_REDUCE_WORKER_POOL_SIZE as usize);
+        let mut worker_threads: Vec<Worker> = Vec::with_capacity(*statics::MAP_REDUCE_WORKER_POOL_SIZE as usize);
 
         for _ in 0..*statics::MAP_REDUCE_WORKER_POOL_SIZE {
-            worker_threads.push(MapReduceWorker::start(
+            worker_threads.push(Worker::start(
                 account_id,
                 contents.clone(),
                 path.clone(),
@@ -156,7 +124,7 @@ impl MapReduceWorkerPool {
                 select_loop! {
                     recv(mapreduce_in_receiver, vertex) => {
                         // If this errors out, all of the workers are dead
-                        if worker_in_sender.send(MapReduceWorkerTask::Map(vertex)).is_err() {
+                        if worker_in_sender.send(WorkerTask::Map(vertex)).is_err() {
                             should_force_shutdown = true;
                         } else {
                             pending_tasks += 1;
@@ -167,7 +135,7 @@ impl MapReduceWorkerPool {
 
                         if let Some(last_reduced_item_inner) = last_reduced_item {
                             // If this errors out, all of the workers are dead
-                            if worker_in_sender.send(MapReduceWorkerTask::Reduce((last_reduced_item_inner, value))).is_err() {
+                            if worker_in_sender.send(WorkerTask::Reduce((last_reduced_item_inner, value))).is_err() {
                                 should_force_shutdown = true;
                             } else {
                                 pending_tasks += 1;
